Update IO_WSSFM10.cpp
Sigfox messages by default are ended with a new line feed. This will make an issue if you want to organize your output in the way you want. E.g. if you want ID and PAC to be in the same line separated by a comma (to be a CSV format).

I recorded a Sigfox response to getID() function with a logic analyzer and this is the result (some parts are covered due to privacy).
https://pasteboard.co/I87o12ZM.png
Here you can find ASCII table as well for all characters, including 0x0A: http://www.asciitable.com/
To solve this issue, a small change is needed that is suggested by this pull request.

// get data
String getData(){
  String data = "";
  char output;

  while (!Sigfox.available()){
     blink();
  }

  while(Sigfox.available()){
    output = Sigfox.read();
    if (output!=0x0A){
      data += output;
    }
    
    delay(10);
  }

  return data;  
}